requirements.txt
python-dotenv
pandas
psycopg2-binary
rapidfuzz
